+++ CHANGELOG.md
@@ -3,10 +3,13 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [1.0.0] - 2023-01-29
 ### Added
+- Animations to the loginar model
 - Vacuum urn upgrade now works. It pulls nearby items towards a placed urn like a magnet, and stores anything that gets close.
 - Italian translation (giok3r)
+### Removed
+- The unimplemented "hotbar swapper upgrade". It may or may not appear in the mod someday.
 ### Changed
 - Increased loginar base health from 20 to 30